Move collection of metrics out of do_action and into Context
