+++ crates/modelardb_server/src/remote/mod.rs
@@ -51,7 +51,6 @@ use modelardb_types::flight::protocol;
 use modelardb_types::functions;
 use modelardb_types::types::{ServerMode, Table, TimeSeriesTableMetadata};
 use prost::Message;
-use sysinfo::{Disks, System};
 use tokio::sync::mpsc::{self, Sender};
 use tokio::task;
 use tokio_stream::wrappers::ReceiverStream;
@@ -290,29 +289,6 @@ pub fn table_name_from_flight_descriptor(
 /// to return.
 fn empty_record_batch_stream() -> SendableRecordBatchStream {
     Box::pin(EmptyRecordBatchStream::new(Arc::new(Schema::empty())))
-}
-
-/// Return the used and total disk space in bytes for the disk holding the local data folder. The
-/// disk is identified by finding the mounted disk whose mount point is the longest prefix of the
-/// local data folder path. If no disk matches, e.g., because the data folder is in memory, the
-/// largest-capacity disk is used instead. If no disks are found, `(0, 0)` is returned.
-fn local_data_folder_disk_space(context: &Context) -> (u64, u64) {
-    let disks = Disks::new_with_refreshed_list();
-    let location = context.data_folders.local_data_folder.location();
-
-    let maybe_disk = disks
-        .iter()
-        .filter(|disk| location.starts_with(&*disk.mount_point().to_string_lossy()))
-        .max_by_key(|disk| disk.mount_point().as_os_str().len())
-        .or_else(|| disks.iter().max_by_key(|disk| disk.total_space()));
-
-    if let Some(disk) = maybe_disk {
-        let total = disk.total_space();
-        let used = total.saturating_sub(disk.available_space());
-        (used, total)
-    } else {
-        (0, 0)
-    }
 }
 
 /// Convert an `error` to a [`Status`] with [`tonic::Code::InvalidArgument`] as the code and `error`
@@ -1120,62 +1096,7 @@ impl FlightService for FlightServiceHandler {
                 })
             }))))
         } else if action.r#type == "NodeMetrics" {
-            let mut system = System::new();
-
-            // Sample the CPU twice, separated by the minimum update interval, since a single
-            // refresh reads zero.
-            system.refresh_cpu_usage();
-            tokio::time::sleep(sysinfo::MINIMUM_CPU_UPDATE_INTERVAL).await;
-            system.refresh_cpu_usage();
-
-            let cpu_usage_percentage = system.global_cpu_usage() as f64;
-            let cpu_count = system.cpus().len() as u32;
-
-            system.refresh_memory();
-            let used_memory_in_bytes = system.used_memory();
-            let total_memory_in_bytes = system.total_memory();
-
-            let (used_disk_space_in_bytes, total_disk_space_in_bytes) =
-                local_data_folder_disk_space(&self.context);
-
-            // Pair each reserved value from the configuration with the remaining value from the
-            // storage engine to compute how much of the reservation is currently in use.
-            let configuration_manager = self.context.configuration_manager.read().await;
-            let storage_engine = self.context.storage_engine.read().await;
-
-            let ingested_reserved_memory_in_bytes =
-                configuration_manager.ingested_reserved_memory_in_bytes();
-            let uncompressed_reserved_memory_in_bytes =
-                configuration_manager.uncompressed_reserved_memory_in_bytes();
-            let compressed_reserved_memory_in_bytes =
-                configuration_manager.compressed_reserved_memory_in_bytes();
-
-            let ingested_used_memory_in_bytes = (ingested_reserved_memory_in_bytes as i64
-                - storage_engine.remaining_ingested_memory_in_bytes())
-            .max(0) as u64;
-            let uncompressed_used_memory_in_bytes = (uncompressed_reserved_memory_in_bytes as i64
-                - storage_engine.remaining_uncompressed_memory_in_bytes())
-            .max(0) as u64;
-            let compressed_used_memory_in_bytes = (compressed_reserved_memory_in_bytes as i64
-                - storage_engine.remaining_compressed_memory_in_bytes())
-            .max(0) as u64;
-
-            let node_metrics = protocol::NodeMetrics {
-                cpu_usage_percentage,
-                cpu_count,
-                used_memory_in_bytes,
-                total_memory_in_bytes,
-                used_disk_space_in_bytes,
-                total_disk_space_in_bytes,
-                ingested_used_memory_in_bytes,
-                ingested_reserved_memory_in_bytes,
-                uncompressed_used_memory_in_bytes,
-                uncompressed_reserved_memory_in_bytes,
-                compressed_used_memory_in_bytes,
-                compressed_reserved_memory_in_bytes,
-            };
-
-            let protobuf_bytes = node_metrics.encode_to_vec();
+            let protobuf_bytes = self.context.node_metrics().await.encode_to_vec();
 
             Ok(Response::new(Box::pin(stream::once(async {
                 Ok(FlightResult {
